Изменение масштаба фото вынесено в отдельный файл change-scale-photo.
