+++ js/form.js
@@ -1,6 +1,7 @@
 import {isEscapeKey} from './utils.js';
 import {createSliderEffects, applyOriginalEffect} from './create-slider-effects.js';
-import {MAX_HASHTAG_SYMBOLS, MAX_COMMENT_SYMBOLS, MAX_QUANTITY_HASHTAGS, DEFAULT_SCALE_PHOTO, STEP_SCALE_PHOTO, MIN_SCALE_VALUE_PHOTO,MAX_SCALE_VALUE_PHOTO, regexp} from './const.js';
+import {MAX_HASHTAG_SYMBOLS, MAX_COMMENT_SYMBOLS, MAX_QUANTITY_HASHTAGS, regexp} from './const.js';
+import {changeScalePhoto, resetScalePhoto} from './change-scale-photo.js';
 
 const pageBody = document.querySelector('body');
 const uploadFormPhoto = pageBody.querySelector('.img-upload__form'); // Находим форму
@@ -10,9 +11,6 @@ const buttonResetUploadFormPhoto = uploadFormPhoto.querySelector('.img-upload__c
 const hashtagsInput = uploadFormPhoto.querySelector('.text__hashtags');// поле для ввода хэштегов
 const commentInput = uploadFormPhoto.querySelector('.text__description'); // поле для ввода комментария
 
-const buttonScaleControlSmaller = uploadFormPhoto.querySelector('.scale__control--smaller');
-const buttonScaleControlBigger = uploadFormPhoto.querySelector('.scale__control--bigger');
-const scaleValueInput = uploadFormPhoto.querySelector('.scale__control--value');
 const uploadPhotoPreview = uploadFormPhoto.querySelector('.img-upload__preview img');
 
 const changeEffectInput = uploadFormPhoto.querySelector('.effect-level__value');// здесь записываем value при движении ползунка.
@@ -24,10 +22,6 @@ const pristine = new Pristine(uploadFormPhoto, {
   errorTextParent: 'img-upload__field-wrapper',
 });
 
-const resetScalePhoto = () => {
-  uploadPhotoPreview.style.transform = 'scale(1)';
-  scaleValueInput.setAttribute('value', `${DEFAULT_SCALE_PHOTO }%`);
-};
 
 const onDocumentKeydown = (evt) => {
   if (isEscapeKey(evt) && document.activeElement !== commentInput && document.activeElement !== hashtagsInput) {
@@ -138,48 +132,7 @@ uploadFormPhoto.addEventListener('submit', (evt) => {
 
 buttonResetUploadFormPhoto.addEventListener('click', closeUploadFormPhoto);
 
-buttonScaleControlSmaller.addEventListener ('click', () => {
-
-  const currentValueInput = parseFloat(scaleValueInput.value.slice(0, -1));
-
-  if (currentValueInput > MIN_SCALE_VALUE_PHOTO) {
-
-    const newScaleValueInput = currentValueInput - STEP_SCALE_PHOTO;
-    scaleValueInput.setAttribute('value', `${newScaleValueInput }%`);
-    scaleValueInput.value = `${newScaleValueInput }%`;
-    const scaleValue = newScaleValueInput / 100;
-    uploadPhotoPreview.style.transform = `scale(${scaleValue})`;
-
-    if (newScaleValueInput === MIN_SCALE_VALUE_PHOTO) {
-      buttonScaleControlSmaller.setAttribute('disabled', '');
-    }
-    if (newScaleValueInput < MAX_SCALE_VALUE_PHOTO) {
-      buttonScaleControlBigger.removeAttribute ('disabled');
-    }
-  }
-});
-
-buttonScaleControlBigger.addEventListener ('click', () => {
-  const currentValueInput = parseFloat(scaleValueInput.value.slice(0, -1));
-
-  if (currentValueInput < MAX_SCALE_VALUE_PHOTO) {
-
-    const newScaleValueInput = currentValueInput + STEP_SCALE_PHOTO;
-    scaleValueInput.setAttribute('value', `${newScaleValueInput }%`);
-    scaleValueInput.value = `${newScaleValueInput }%`;
-    const scaleValue = newScaleValueInput / 100;
-    uploadPhotoPreview.style.transform = `scale(${scaleValue})`;
-
-    if (newScaleValueInput > MIN_SCALE_VALUE_PHOTO) {
-      buttonScaleControlSmaller.removeAttribute('disabled');
-      if (newScaleValueInput === MAX_SCALE_VALUE_PHOTO) {
-        buttonScaleControlBigger.setAttribute('disabled', '');
-      }
-    }
-  }
-
-});
-
 createSliderEffects ();
+changeScalePhoto ();
 
 export {openUploadFormPhoto};
